refactor: 스프링 버전 변경에 따른 엔티티 수정 및 파일 변경
