+++ member-service/src/main/java/com/blockpage/memberservice/adaptor/infrastructure/entity/EmotionEntity.java
@@ -1,13 +1,13 @@
 package com.blockpage.memberservice.adaptor.infrastructure.entity;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -30,7 +30,7 @@ public class EmotionEntity {
     private MemberEntity memberEntity;
 
     @Column
-    private  Long episodeId;
+    private Long episodeId;
 
     @Column
     private Long commentId;
