+++ member-service/src/main/java/com/blockpage/memberservice/adaptor/infrastructure/entity/MemberEntity.java
@@ -1,7 +1,13 @@
 package com.blockpage.memberservice.adaptor.infrastructure.entity;
 
 import com.blockpage.memberservice.adaptor.infrastructure.view.Role;
-import jakarta.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -17,6 +23,9 @@ public class MemberEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    @Column
+    private Long kakaoId;
 
     @Column
     private String email;
