+++ member-service/src/main/java/com/blockpage/memberservice/adaptor/infrastructure/repository/InterestRepository.java
@@ -4,6 +4,7 @@ import com.blockpage.memberservice.adaptor.infrastructure.entity.InterestEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface InterestRepository extends JpaRepository<InterestEntity,Long> {
+public interface InterestRepository extends JpaRepository<InterestEntity, Long> {
+
     List<InterestEntity> findAllByMemberEntityId(Long memberId);
 }
